enhance tests
