+++ tests/__main__.py
@@ -1,0 +1,8 @@
+import unittest
+from api_test import TestApi
+from build_test import TestBuild
+from project_test import TestProject
+from requester_test import TestRequester
+
+if __name__ == "__main__":
+    unittest.main()
