+++ tests/build_test.py
@@ -1,0 +1,10 @@
+import unittest
+import requests_mock
+
+
+class TestBuild(unittest.TestCase):
+    def test_from_json(self):
+        pass # Test not implemented yet
+
+if __name__ == "__main__":
+    unittest.main()
